feat(diesel): remove most hard-coded table definitions, add migrations
In anticipation of https://github.com/diesel-rs/diesel/issues/43, this
proactivey removes the hand-coded table definitions while moving the
views to a new module called `view`. The next step in this work is to
refactor the Rocket routes to permit returning either JSON or a CSV on
most/all API endpoints and then refactor the table names and
relationships so that the overall codebase is more approachable and
maintainable. After the refactor, we can then re-code the
optional migrations to fit the new platform.

By separating out the hand-coded views from the automatically-generated
table definitions, once diesel-rs/diesel#43 is closed, we will be able
to remove the `view` module entirely. By divorcing views from tables, we
also stand to benefit in that removing the `structs` module will be
significantly easier.